updated CourseService

+++ src/main/java/com/example/wbdvf19jadaserverjava/services/CourseService.java
@@ -12,16 +12,16 @@ public class CourseService {
     List<Course> courses = new ArrayList<Course>();
 
     {
-        courses.add(new Course(123, "CS 5610"));
-        courses.add(new Course(234,"CS 4500"));
-        courses.add(new Course(345, "CS 4550"));
-        courses.add(new Course(456,"CS 3600"));
-        courses.add(new Course(567, "CS 3650"));
+        courses.add(new Course(123L, "CS 5610"));
+        courses.add(new Course(234L,"CS 4500"));
+        courses.add(new Course(345L, "CS 4550"));
+        courses.add(new Course(456L,"CS 3600"));
+        courses.add(new Course(567L, "CS 3650"));
     }
 
     @PutMapping("/api/courses/{courseId}")
     public List<Course> updateCourse(
-            @PathVariable("courseId") int id,
+            @PathVariable("courseId") long id,
             @RequestBody Course newCourse) {
         for (Course course : courses) {
             if (course.getId() == id) {
@@ -45,7 +45,7 @@ public class CourseService {
 
     @GetMapping("/api/courses/{courseId}")
     public Course findCourseById(
-            @PathVariable("courseId") int id) {
+            @PathVariable("courseId") long id) {
         for (Course course : courses) {
             if (course.getId() == id) {
                 return course;
@@ -55,7 +55,7 @@ public class CourseService {
     }
 
     @DeleteMapping("/api/courses/{courseId}")
-    public List<Course> deleteCourse(@PathVariable("courseId") int cor) {
+    public List<Course> deleteCourse(@PathVariable("courseId") long cor) {
         courses = courses
                 .stream()
                 .filter(course -> !(course.getId() == cor))
